JSONObject constructor with custom map supplier
Added a constructor for JSONObject allowing the user to choose the Map implementation (instead of the default HashMap).
By choice of map implementations like LinkedHashMap or some sorted map the user can influence the iteration order and thus especially the order of the toString-methods' output. While this is not important for machine-readable JSON, for a pretty-printed, human-readable JSON created by #toString(int) (probably meant for humans to edit the data), a custom order may increase readability and maintainability of the generated code.