Create module-info.java
When using this library in a modular app you need to add to `module-info.java`:
```
requires hjson;
```
Which will raise a warning, e.g. `Name of automatic module 'hjson' is unstable, it is derived from the module's file name`, as obviously it is an automatic module. By adding a `module-info.java` we formalize the name as `hjson` (which makes it backward compatible with current automatic module resolution).

More on automatic modules: https://stackoverflow.com/a/46742802 correctly, this should be all that's needed in order for hjson