Added new constructor to keep items ordered
Construct an empty JSONObject specifying if the items are ordered based on the order of insertion. If ordered, a LinkedHashMap is used instead of a HashMap. Even if the JSON standard does not mandate the ordering of properties, in some cases it is requested by the specific use case. For example the ordering of the items is requested by <a href="https://spec.graphql.org/June2018/#sec-Serialized-Map-Ordering">GraphQL specification</a>. The default implementation should use the HashMap because if the order is not requested, it is more efficient using the HashMap in terms of memory consumption and runtime cost for most operations.