do not return 'null' for toString
Motivation:

`null` is not a valid return value for Object#toString; for example, see:

https://stackoverflow.com/questions/32378264/when-we-override-the-tostring-method-we-should-always-return-a-string-represen

Additionally, catching Exception means that any RuntimeException (which
is likely a bug in the underlying code) will be supressed, making it
harder to figure out what is going wrong.

Modification:

The JSONObject#toString contract requires toString to return valid JSON.

The code could return a valid JSON by returning a textural description
of the JSONException as a quoted string; however, this would not make it
clear than an error has occurred, and would likely be incompatible with
the expected structure of the JSON (the schema).

As JSONException is a RuntimeException, the code simply allow it to
propagate.

Result:

JSONObject#toString no longer returns `null` under exceptional
circumstances.